KDKPAS-3439 Make `get-report` output cleaner
Previously when the polling finished, the output would be a bit messy,
as part of the previous message would get overwritten by the next line,
causing it to look something like
"Polling is done. Transfer has been processed. the DPS ingest for the SIP validation report...|"

Print additional newline so we don't partially overwrite the line.